Refactor GroceryPanel to use custom hook for data fetching and simplify state management
